Client: Removed some unused variables

+++ Orion+#/Source/Client/Source/Modules/C_General.cs
@@ -144,17 +144,17 @@ namespace Engine
 			C_Variables.Ping = -1;
 			
 			// set values for directional blocking arrows
-			C_Variables.DirArrowX[1] = (byte) 12; // up
-			C_Variables.DirArrowY[1] = (byte) 0;
+			C_Variables.DirArrowX[1] = 12; // up
+			C_Variables.DirArrowY[1] = 0;
 
-			C_Variables.DirArrowX[2] = (byte) 12; // down
-			C_Variables.DirArrowY[2] = (byte) 23;
+			C_Variables.DirArrowX[2] = 12; // down
+			C_Variables.DirArrowY[2] = 23;
 
-			C_Variables.DirArrowX[3] = (byte) 0; // left
-			C_Variables.DirArrowY[3] = (byte) 12;
+			C_Variables.DirArrowX[3] = 0; // left
+			C_Variables.DirArrowY[3] = 12;
 
-			C_Variables.DirArrowX[4] = (byte) 23; // right
-			C_Variables.DirArrowY[4] = (byte) 12;
+			C_Variables.DirArrowX[4] = 23; // right
+			C_Variables.DirArrowY[4] = 12;
 			
 			//set gui switches
 			C_UpdateUI.HudVisible = true;
